Reduced code duplication by combining these three very similar tests into a single file.
